// src/globals/environments/enums/keys.ts
export enum Keys {
  MICROSERVICIO = 'MICROSERVICIO',
  PORT = 'PORT',
  HOST = 'HOST',
  DATABASE_HOST = 'DATABASE_HOST',
  DATABASE_NAME = 'DATABASE_NAME',
  DATABASE_PORT = 'DATABASE_PORT',
  DATABASE_USER = 'DATABASE_USER',
  DATABASE_PASSWORD = 'DATABASE_PASSWORD',
  DATABASE_CONNECTION_LIMIT = 'DATABASE_CONNECTION_LIMIT',
  TYPEORM_CONNECTION = 'TYPEORM_CONNECTION',
  TYPEORM_HOST = 'TYPEORM_HOST',
  TYPEORM_USERNAME = 'TYPEORM_USERNAME',
  TYPEORM_PASSWORD = 'TYPEORM_PASSWORD',
  TYPEORM_DATABASE = 'TYPEORM_DATABASE',
  TYPEORM_PORT = 'TYPEORM_PORT',
  TYPEORM_LOGGING = 'TYPEORM_LOGGING',
  TYPEORM_SYNCHRONIZE = 'TYPEORM_SYNCHRONIZE',
  TYPEORM_ENTITIES = 'TYPEORM_ENTITIES',
  TYPEORM_MIGRATIONS = 'TYPEORM_MIGRATIONS',
  TYPEORM_MIGRATIONS_DIR = 'TYPEORM_MIGRATIONS_DIR',
  REDIS_NAME = 'REDIS_NAME',
  REDIS_HOST = 'REDIS_HOST',
  REDIS_PORT = 'REDIS_PORT',
  REDIS_PASSWORD = 'REDIS_PASSWORD',
}